+++ src/components.d.ts
@@ -28,6 +28,29 @@ export namespace Components {
     */
     'configSettings'?: {[packageName:string]:Object};
   }
+  interface PieAuthorAttributes extends StencilHTMLAttributes {
+    /**
+    * Adds a preview view which will render the content in another tab as it may appear to a student or instructor.
+    */
+    'addPreview'?: boolean;
+    /**
+    * The Pie config model.
+    */
+    'config'?: ItemConfig;
+    /**
+    * To customize the standard behaviour provided by interaction configuration views you can  provide settings key-ed by the package name.  e.g.  `{ '@pie-element/inline-choice': { promptLabel: 'Item Stem' } }`  The settings that are configurable for each authoring view are documented in  the `@package-name/docs` folder for each package.
+    */
+    'configSettings'?: {[packageName:string]:Object};
+    /**
+    * Emmitted when the content models in the config have ben set on the content
+    */
+    'onModelLoaded'?: (event: CustomEvent) => void;
+    /**
+    * Emmitted when the model for the content has been updated within the ui due to user action.
+    */
+    'onModelUpdated'?: (event: CustomEvent) => void;
+  }
+
   interface PieLoader {
     /**
     * Loads the custom elments defined by the PIEs, if they are not already loaded.